1286 - add alert max-width to match the PaymentSummary alert
